+++ api/controllers/shop/CheckoutController.php
@@ -4,7 +4,7 @@ namespace api\controllers\shop;
 
 use shop\cart\Cart;
 use shop\forms\Shop\Order\OrderForm;
-use shop\services\Shop\OrderService;
+use shop\useCases\Shop\OrderService;
 use Yii;
 use yii\helpers\Url;
 use yii\rest\Controller;
