+++ backend/controllers/blog/PostController.php
@@ -3,7 +3,7 @@
 namespace backend\controllers\blog;
 
 use shop\forms\manage\Blog\Post\PostForm;
-use shop\services\manage\Blog\PostManageService;
+use shop\useCases\manage\Blog\PostManageService;
 use Yii;
 use shop\entities\Blog\Post\Post;
 use backend\forms\Blog\PostSearch;
