+++ backend/controllers/shop/OrderController.php
@@ -4,7 +4,7 @@ namespace backend\controllers\shop;
 
 use shop\forms\manage\Shop\Order\OrderEditForm;
 use shop\forms\manage\Shop\OrderForm;
-use shop\services\manage\Shop\OrderManageService;
+use shop\useCases\manage\Shop\OrderManageService;
 use Yii;
 use shop\entities\Shop\Order\Order;
 use backend\forms\Shop\OrderSearch;
