+++ console/controllers/RoleController.php
@@ -3,7 +3,7 @@
 namespace console\controllers;
 
 use shop\entities\User\User;
-use shop\services\manage\UserManageService;
+use shop\useCases\manage\UserManageService;
 use Yii;
 use yii\console\Controller;
 use yii\console\Exception;
