+++ frontend/controllers/ContactController.php
@@ -1,7 +1,7 @@
 <?php
 namespace frontend\controllers;
 
-use shop\services\ContactService;
+use shop\useCases\ContactService;
 use Yii;
 use yii\web\Controller;
 use shop\forms\ContactForm;
