+++ frontend/controllers/auth/AuthController.php
@@ -2,7 +2,7 @@
 namespace frontend\controllers\auth;
 
 use common\auth\Identity;
-use shop\services\auth\AuthService;
+use shop\useCases\auth\AuthService;
 use Yii;
 use yii\web\Controller;
 use shop\forms\auth\LoginForm;
