+++ frontend/controllers/auth/ResetController.php
@@ -1,7 +1,7 @@
 <?php
 namespace frontend\controllers\auth;
 
-use shop\services\auth\PasswordResetService;
+use shop\useCases\auth\PasswordResetService;
 use Yii;
 use yii\web\BadRequestHttpException;
 use yii\web\Controller;
