+++ frontend/controllers/cabinet/NetworkController.php
@@ -2,7 +2,7 @@
 
 namespace frontend\controllers\cabinet;
 
-use shop\services\auth\NetworkService;
+use shop\useCases\auth\NetworkService;
 use Yii;
 use yii\authclient\ClientInterface;
 use yii\helpers\ArrayHelper;
